+++ socialhome/search/tests/test_views.py
@@ -1,11 +1,16 @@
+from unittest.mock import patch
+from uuid import uuid4
+
 import haystack
 from django.contrib.auth.models import AnonymousUser
 from django.test import RequestFactory
 from django.urls import reverse
+from federation.entities import base
 
 from socialhome.enums import Visibility
 from socialhome.search.views import GlobalSearchView
 from socialhome.tests.utils import SocialhomeCBVTestCase, SocialhomeTestCase
+from socialhome.users.models import Profile
 from socialhome.users.tests.factories import ProfileFactory, UserFactory
 from socialhome.users.views import ProfileAllContentView
 
@@ -70,7 +75,22 @@ class TestGlobalSearchView(SocialhomeTestCase):
         self.assertResponseContains(self.public_profile.username_part, html=False)
 
     def test_direct_profile_match_goes_to_profile_view(self):
-        self.get("%s?q=%s" % (reverse("search:global"), self.public_profile.handle))
+        self.get("%s?q=%s" % (reverse("search:global"), self.public_profile.handle), follow=True)
         self.assertResponseNotContains("Results", html=False)
         self.assertEqual(self.context["view"].__class__, ProfileAllContentView)
         self.assertEqual(self.context["object"], self.public_profile)
+
+    @patch("socialhome.search.views.retrieve_remote_profile")
+    def test_search_by_handle_fetches_unknown_profile(self, mock_retrieve):
+        guid = str(uuid4())
+        mock_retrieve.return_value = base.Profile(
+            name="I don't exist locally",
+            handle="i-dont-exist-locally@example.com",
+            guid=guid,
+            public=True,
+        )
+        self.get("%s?q=%s" % (reverse("search:global"), "i-dont-exist-locally@example.com"), follow=True)
+        profile = Profile.objects.get(guid=guid)
+        self.assertResponseNotContains("Results", html=False)
+        self.assertEqual(self.context["view"].__class__, ProfileAllContentView)
+        self.assertEqual(self.context["object"], profile)