+++ docs/changelog.rst
@@ -9,11 +9,13 @@ Changelog
 Added
 .....
 
-* Added search for profiles
+* Added search for profiles (`#163 <https://github.com/jaywink/socialhome/issues/163>`_)
 
   There is now a global search in the right side of the header. The search returns matches for local and remote profiles based on their name and username part of the handle. Profiles marked with visibility ``Self`` or ``Limited`` are excluded from the search results. Profiles marked with visibility ``Site`` will be excluded if not logged in, leaving only public profile results. If a direct match happens with a full handle, a redirect is done directly to the searched profile.
 
   **IMPORTANT for node maintainers**. After pulling in this change, you MUST run the command ``python manage.py rebuild_index`` to create the search index. Not doing this will cause an error to be raised when trying to search. The indexes are kept up to date automatically after running this command once.
+
+* When searching for profiles based on handle, fetch profile from remote if it isn't found locally (`#163 <https://github.com/jaywink/socialhome/issues/163>`_)
 
 0.1.0 (2017-07-27)
 ------------------